[Use Map Editor #12] remove generated tileset from VCS and target temporary files with disabled tileset generator test because I don't wanna use it anymore. Initially I thought I should use tiles to create nice maps where the visuals can be previewed from. I disregarded by own opinion because I considered the recommendations in the documentation, the examples shown everywhere and what the editor was designed to do. That was a very bad decision considering my needs because tiling just doesn't work the way I hoped it to work like.
Tiles suck for dynamically formed stuff, suck for one-layer high stuff, suck for adding metadata, suck for dynamic stuff, suck for moving stuff around, just suck in terms of flexibility. They just don't suck for displaying visuals of the map and creating nice terrains. Terrain tooling also sucks for creating any single object not in the form of a rectangle.

I'll use objects instead. It will be way more flexible and enable me to create custom forms in the editor too which can be used very well for a physics-focused game.
Damn that generator was a waste of time.